A php script is written to create a table in postgreSQL database.
A php script is written to create a table in postgreSQL database. This script creates the table and inserts the data from a CSV file named users.csv from the local folder and updates in the postgreSQL database. Validations are done according to the document.
Name and surname fields are capitalised. Email is validated and converted to lowercase and inserted.